+++ menu/colorize.rb
@@ -1,13 +1,9 @@
-class App
+class Colorizer
   def colorize_output(color_code, statements)
     puts "\e[#{color_code}m#{statements}\e[0m"
   end
 
   def colorize_outprint(color_code, statements)
     print "\e[#{color_code}m#{statements}\e[0m"
-  end
-
-  def app_work
-    puts 'app is working fine'
   end
 end
