+++ menu/games_operations.rb
